fixes, linting
