+++ src/paperap/resources/base.py
@@ -529,300 +529,6 @@ class StandardResource(BaseResource[_StandardModel, _StandardQuerySet]):
         data = self.transform_data_output(**data)
         return self.update_dict(model.id, **data)
 
-    def bulk_action(self, action: str, ids: list[int], **kwargs: Any) -> dict[str, Any]:
-        """
-        Perform a bulk action on multiple resources.
-
-        Args:
-            action: The action to perform (e.g., "delete", "update", etc.)
-            ids: List of resource IDs to perform the action on
-            **kwargs: Additional parameters for the action
-
-        Returns:
-            The API response
-
-        Raises:
-            ConfigurationError: If the bulk endpoint is not defined
-
-        """
-        # Signal before bulk action
-        signal_params = {"resource": self.name, "action": action, "ids": ids, "kwargs": kwargs}
-        registry.emit("resource.bulk_action:before", "Emitted before bulk action", args=[self], kwargs=signal_params)
-
-        # Prepare the data for the bulk action
-        data = {"method": action, "documents": ids, "parameters": kwargs}
-
-        if not (url := self.get_endpoint("bulk_edit", resource=self.name)):
-            raise ConfigurationError(f"Bulk edit endpoint not defined for resource {self.name}")
-
-        response = self.client.request("POST", url, data=data)
-
-        # Signal after bulk action
-        registry.emit(
-            "resource.bulk_action:after",
-            "Emitted after bulk action",
-            args=[self],
-            kwargs={**signal_params, "response": response},
-        )
-
-        return response or {}
-
-    def bulk_delete(self, ids: list[int]) -> dict[str, Any]:
-        """
-        Delete multiple resources at once.
-
-        Args:
-            ids: List of resource IDs to delete
-
-        Returns:
-            The API response
-
-        """
-        return self.bulk_action("delete", ids)
-
-    def bulk_update(self, ids: list[int], **kwargs: Any) -> dict[str, Any]:
-        """
-        Update multiple resources at once.
-
-        This is a generic method that transforms the data before sending.
-        For document-specific operations, use the specialized methods.
-
-        Args:
-            ids: List of resource IDs to update
-            **kwargs: Fields to update on all resources
-
-        Returns:
-            The API response
-
-        """
-        # Transform the data before sending
-        data = self.transform_data_output(**kwargs)
-        return self.bulk_action("update", ids, **data)
-
-    def bulk_reprocess(self, ids: list[int]) -> dict[str, Any]:
-        """
-        Reprocess multiple documents.
-
-        Args:
-            ids: List of document IDs to reprocess
-
-        Returns:
-            The API response
-
-        """
-        return self.bulk_action("reprocess", ids)
-
-    def bulk_merge(
-        self, ids: list[int], metadata_document_id: int | None = None, delete_originals: bool = False
-    ) -> dict[str, Any]:
-        """
-        Merge multiple documents.
-
-        Args:
-            ids: List of document IDs to merge
-            metadata_document_id: Apply metadata from this document to the merged document
-            delete_originals: Whether to delete the original documents after merging
-
-        Returns:
-            The API response
-
-        """
-        params = {}
-        if metadata_document_id is not None:
-            params["metadata_document_id"] = metadata_document_id
-        if delete_originals:
-            params["delete_originals"] = True
-
-        return self.bulk_action("merge", ids, **params)
-
-    def bulk_split(self, document_id: int, pages: list[int], delete_originals: bool = False) -> dict[str, Any]:
-        """
-        Split a document.
-
-        Args:
-            document_id: Document ID to split
-            pages: List of pages to split
-            delete_originals: Whether to delete the original document after splitting
-
-        Returns:
-            The API response
-
-        """
-        params = {"pages": pages}
-        if delete_originals:
-            params["delete_originals"] = True
-
-        return self.bulk_action("split", [document_id], **params)
-
-    def bulk_rotate(self, ids: list[int], degrees: int) -> dict[str, Any]:
-        """
-        Rotate documents.
-
-        Args:
-            ids: List of document IDs to rotate
-            degrees: Degrees to rotate (must be 90, 180, or 270)
-
-        Returns:
-            The API response
-
-        """
-        if degrees not in (90, 180, 270):
-            raise ValueError("Degrees must be 90, 180, or 270")
-
-        return self.bulk_action("rotate", ids, degrees=degrees)
-
-    def bulk_delete_pages(self, document_id: int, pages: list[int]) -> dict[str, Any]:
-        """
-        Delete pages from a document.
-
-        Args:
-            document_id: Document ID
-            pages: List of page numbers to delete
-
-        Returns:
-            The API response
-
-        """
-        return self.bulk_action("delete_pages", [document_id], pages=pages)
-
-    def bulk_modify_custom_fields(
-        self, ids: list[int], add_custom_fields: dict[int, Any] | None = None, remove_custom_fields: list[int] | None = None
-    ) -> dict[str, Any]:
-        """
-        Modify custom fields on multiple documents.
-
-        Args:
-            ids: List of document IDs to update
-            add_custom_fields: Dictionary of custom field ID to value pairs to add
-            remove_custom_fields: List of custom field IDs to remove
-
-        Returns:
-            The API response
-
-        """
-        params = {}
-        if add_custom_fields:
-            params["add_custom_fields"] = add_custom_fields
-        if remove_custom_fields:
-            params["remove_custom_fields"] = remove_custom_fields
-
-        return self.bulk_action("modify_custom_fields", ids, **params)
-
-    def bulk_modify_tags(self, ids: list[int], add_tags: list[int] = None, remove_tags: list[int] = None) -> dict[str, Any]:
-        """
-        Modify tags on multiple documents.
-
-        Args:
-            ids: List of document IDs to update
-            add_tags: List of tag IDs to add
-            remove_tags: List of tag IDs to remove
-
-        Returns:
-            The API response
-
-        """
-        params = {}
-        if add_tags:
-            params["add_tags"] = add_tags
-        if remove_tags:
-            params["remove_tags"] = remove_tags
-
-        return self.bulk_action("modify_tags", ids, **params)
-
-    def bulk_add_tag(self, ids: list[int], tag_id: int) -> dict[str, Any]:
-        """
-        Add a tag to multiple documents.
-
-        Args:
-            ids: List of document IDs to update
-            tag_id: Tag ID to add
-
-        Returns:
-            The API response
-
-        """
-        return self.bulk_action("add_tag", ids, tag=tag_id)
-
-    def bulk_remove_tag(self, ids: list[int], tag_id: int) -> dict[str, Any]:
-        """
-        Remove a tag from multiple documents.
-
-        Args:
-            ids: List of document IDs to update
-            tag_id: Tag ID to remove
-
-        Returns:
-            The API response
-
-        """
-        return self.bulk_action("remove_tag", ids, tag=tag_id)
-
-    def bulk_set_correspondent(self, ids: list[int], correspondent_id: int) -> dict[str, Any]:
-        """
-        Set correspondent for multiple documents.
-
-        Args:
-            ids: List of document IDs to update
-            correspondent_id: Correspondent ID to assign
-
-        Returns:
-            The API response
-
-        """
-        return self.bulk_action("set_correspondent", ids, correspondent=correspondent_id)
-
-    def bulk_set_document_type(self, ids: list[int], document_type_id: int) -> dict[str, Any]:
-        """
-        Set document type for multiple documents.
-
-        Args:
-            ids: List of document IDs to update
-            document_type_id: Document type ID to assign
-
-        Returns:
-            The API response
-
-        """
-        return self.bulk_action("set_document_type", ids, document_type=document_type_id)
-
-    def bulk_set_storage_path(self, ids: list[int], storage_path_id: int) -> dict[str, Any]:
-        """
-        Set storage path for multiple documents.
-
-        Args:
-            ids: List of document IDs to update
-            storage_path_id: Storage path ID to assign
-
-        Returns:
-            The API response
-
-        """
-        return self.bulk_action("set_storage_path", ids, storage_path=storage_path_id)
-
-    def bulk_set_permissions(
-        self, ids: list[int], permissions: dict[str, Any] = None, owner_id: int = None, merge: bool = False
-    ) -> dict[str, Any]:
-        """
-        Set permissions for multiple documents.
-
-        Args:
-            ids: List of document IDs to update
-            permissions: Permissions object
-            owner_id: Owner ID to assign
-            merge: Whether to merge with existing permissions (True) or replace them (False)
-
-        Returns:
-            The API response
-
-        """
-        params = {"merge": merge}
-        if permissions:
-            params["set_permissions"] = permissions
-        if owner_id is not None:
-            params["owner"] = owner_id
-
-        return self.bulk_action("set_permissions", ids, **params)
-
     @override
     def delete(self, model_id: int | _StandardModel) -> None:
         """
@@ -848,56 +554,6 @@ class StandardResource(BaseResource[_StandardModel, _StandardQuerySet]):
 
         # Signal after deleting resource
         registry.emit("resource.delete:after", "Emitted after deleting a resource", args=[self], kwargs=signal_params)
-
-    def bulk_edit_objects(
-        self,
-        object_type: str,
-        ids: list[int],
-        operation: str,
-        permissions: dict[str, Any] = None,
-        owner_id: int = None,
-        merge: bool = False
-    ) -> dict[str, Any]:
-        """
-        Bulk edit non-document objects (tags, correspondents, document types, storage paths).
-        
-        Args:
-            object_type: Type of objects to edit ('tags', 'correspondents', 'document_types', 'storage_paths')
-            ids: List of object IDs to edit
-            operation: Operation to perform ('set_permissions' or 'delete')
-            permissions: Permissions object for 'set_permissions' operation
-            owner_id: Owner ID to assign
-            merge: Whether to merge permissions with existing ones (True) or replace them (False)
-            
-        Returns:
-            The API response
-            
-        Raises:
-            ValueError: If operation is not valid
-            ConfigurationError: If the bulk edit endpoint is not defined
-
-        """
-        if operation not in ('set_permissions', 'delete'):
-            raise ValueError(f"Invalid operation '{operation}'. Must be 'set_permissions' or 'delete'")
-
-        data = {
-            "objects": ids,
-            "object_type": object_type,
-            "operation": operation,
-            "merge": merge
-        }
-
-        if permissions:
-            data["permissions"] = permissions
-        if owner_id is not None:
-            data["owner"] = owner_id
-
-        # Use the special endpoint for bulk editing objects
-        url = HttpUrl(f"{self.client.base_url}/api/bulk_edit_objects/")
-
-        response = self.client.request("POST", url, data=data)
-
-        return response or {}
 
     @override
     def update_dict(self, model_id: int, **data: dict[str, Any]) -> _StandardModel:
@@ -936,3 +592,79 @@ class StandardResource(BaseResource[_StandardModel, _StandardQuerySet]):
         )
 
         return model
+
+class BulkEditing:
+
+    def bulk_edit_objects(
+        self : BaseResource, # type: ignore
+        object_type: str,
+        ids: list[int],
+        operation: str,
+        permissions: dict[str, Any] = None,
+        owner_id: int = None,
+        merge: bool = False
+    ) -> dict[str, Any]:
+        """
+        Bulk edit non-document objects (tags, correspondents, document types, storage paths).
+
+        Args:
+            object_type: Type of objects to edit ('tags', 'correspondents', 'document_types', 'storage_paths')
+            ids: List of object IDs to edit
+            operation: Operation to perform ('set_permissions' or 'delete')
+            permissions: Permissions object for 'set_permissions' operation
+            owner_id: Owner ID to assign
+            merge: Whether to merge permissions with existing ones (True) or replace them (False)
+
+        Returns:
+            The API response
+
+        Raises:
+            ValueError: If operation is not valid
+            ConfigurationError: If the bulk edit endpoint is not defined
+
+        """
+        if operation not in ('set_permissions', 'delete'):
+            raise ValueError(f"Invalid operation '{operation}'. Must be 'set_permissions' or 'delete'")
+
+        # Signal before bulk action
+        signal_params = {
+            "object_type": object_type,
+            "operation": operation,
+            "ids": ids,
+            "permissions": permissions,
+            "owner_id": owner_id,
+            "merge": merge
+        }
+        registry.emit(
+            "resource.bulk_edit_objects:before",
+            "Emitted before bulk edit objects",
+            args=[self],
+            kwargs=signal_params,
+        )
+
+        data : dict[str, Any] = {
+            "objects": ids,
+            "object_type": object_type,
+            "operation": operation,
+            "merge": merge
+        }
+
+        if permissions:
+            data["permissions"] = permissions
+        if owner_id is not None:
+            data["owner"] = owner_id
+
+        # Use the special endpoint for bulk editing objects
+        url = HttpUrl(f"{self.client.base_url}/api/bulk_edit_objects/")
+
+        response = self.client.request("POST", url, data=data)
+
+        # Signal after bulk action
+        registry.emit(
+            "resource.bulk_edit_objects:after",
+            "Emitted after bulk edit objects",
+            args=[self],
+            kwargs={**signal_params, "response": response},
+        )
+
+        return response or {}
