+++ tests/unit/resources/test_custom_fields.py
@@ -8,7 +8,7 @@
        Created: 2025-03-21
         Version: 0.0.9
        Author:  Claude
-       Email:   
+       Email:
         Copyright (c) 2025 Jess Mann
 
 ----------------------------------------------------------------------------
@@ -35,7 +35,7 @@ class TestCustomFieldResource(unittest.TestCase):
     def setUp(self) -> None:
         """
         Set up test fixtures.
-        
+
         Written By claude
         """
         self.mock_client = MagicMock(spec=PaperlessClient)
@@ -44,7 +44,7 @@ class TestCustomFieldResource(unittest.TestCase):
     def test_initialization(self) -> None:
         """
         Test that the resource initializes with correct attributes.
-        
+
         Written By claude
         """
         self.assertEqual(self.resource.name, "custom_fields")
@@ -55,7 +55,7 @@ class TestCustomFieldResource(unittest.TestCase):
     def test_all(self) -> None:
         """
         Test that the all() method returns a CustomFieldQuerySet.
-        
+
         Written By claude
         """
         queryset = self.resource.all()
@@ -65,7 +65,7 @@ class TestCustomFieldResource(unittest.TestCase):
     def test_filter(self) -> None:
         """
         Test that the filter() method returns a filtered CustomFieldQuerySet.
-        
+
         Written By claude
         """
         queryset = self.resource.filter(name="Test")
@@ -77,18 +77,18 @@ class TestCustomFieldResource(unittest.TestCase):
     def test_get(self, mock_parse) -> None:
         """
         Test that the get() method fetches a custom field by ID.
-        
+
         Written By claude
         """
         mock_response = {"id": 1, "name": "Test Field"}
         self.mock_client.request.return_value = mock_response
         mock_parse.return_value = CustomField(id=1, name="Test Field")
-        
+
         result = self.resource.get(1)
-        
+
         self.mock_client.request.assert_called_once_with(
-            "GET", 
-            f"{self.resource.name}/1/", 
+            "GET",
+            f"{self.resource.name}/1/",
             params=None
         )
         mock_parse.assert_called_once_with(mock_response)
@@ -99,18 +99,18 @@ class TestCustomFieldResource(unittest.TestCase):
     def test_create(self, mock_parse) -> None:
         """
         Test that the create() method creates a new custom field.
-        
+
         Written By claude
         """
         mock_response = {"id": 1, "name": "New Field"}
         self.mock_client.request.return_value = mock_response
         mock_parse.return_value = CustomField(id=1, name="New Field")
-        
+
         result = self.resource.create(name="New Field")
-        
+
         self.mock_client.request.assert_called_once_with(
-            "POST", 
-            self.resource.name + "/", 
+            "POST",
+            self.resource.name + "/",
             data={"name": "New Field"}
         )
         mock_parse.assert_called_once_with(mock_response)
@@ -120,12 +120,12 @@ class TestCustomFieldResource(unittest.TestCase):
     def test_parse_to_model(self) -> None:
         """
         Test that parse_to_model correctly converts API response to a CustomField.
-        
+
         Written By claude
         """
         data = {"id": 1, "name": "Test Field"}
         result = self.resource.parse_to_model(data)
-        
+
         self.assertIsInstance(result, CustomField)
         self.assertEqual(result.id, 1)
         self.assertEqual(result.name, "Test Field")
@@ -134,24 +134,24 @@ class TestCustomFieldResource(unittest.TestCase):
     def test_update_model(self, mock_update) -> None:
         """
         Test that a model can be updated through the resource.
-        
+
         Written By claude
         """
         model = CustomField(id=1, name="Old Name")
-        
+
         # Mock the client's request method
         mock_response = {"id": 1, "name": "New Name"}
         self.mock_client.request.return_value = mock_response
-        
+
         # Set the client on the model
         model._client = self.mock_client
-        
+
         # Update the model
         model.update(name="New Name")
-        
+
         # Verify the update method was called
         mock_update.assert_called_once()
-        
+
         # Verify the client's request method was called with the correct arguments
         self.mock_client.request.assert_called_once()
 
@@ -159,15 +159,15 @@ class TestCustomFieldResource(unittest.TestCase):
     def test_queryset_name_filter(self, mock_name) -> None:
         """
         Test that the queryset's name filter method is called correctly.
-        
+
         Written By claude
         """
         mock_queryset = MagicMock(spec=CustomFieldQuerySet)
         mock_name.return_value = mock_queryset
-        
+
         queryset = self.resource.all()
         result = queryset.name("Test")
-        
+
         mock_name.assert_called_once_with("Test", exact=True, case_insensitive=True)
         self.assertEqual(result, mock_queryset)
