+++ tests/unit/resources/test_document_download.py
@@ -38,7 +38,7 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def setUp(self):
         """
         Written By claude
-        
+
         Set up test fixtures before each test method.
         Creates a mock client and initializes the resource.
         """
@@ -48,7 +48,7 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_initialization(self):
         """
         Written By claude
-        
+
         Test that the resource is initialized correctly with the right attributes.
         """
         self.assertEqual(self.resource.model_class, DownloadedDocument)
@@ -65,12 +65,12 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_load_download_mode(self):
         """
         Written By claude
-        
+
         Test loading a document in download mode.
         """
         # Create a mock document
         doc = DownloadedDocument(id=1, mode=RetrieveFileMode.DOWNLOAD, original=False)
-        
+
         # Mock the response
         mock_response = MagicMock()
         mock_response.content = b"file content"
@@ -79,15 +79,15 @@ class TestDownloadedDocumentResource(unittest.TestCase):
             "Content-Disposition": 'attachment; filename="test.pdf"',
         }
         self.mock_client.request_raw.return_value = mock_response
-        
+
         # Call the method
         self.resource.load(doc)
-        
+
         # Verify the client was called correctly
         self.mock_client.request_raw.assert_called_once_with(
             "GET", URLS.download, params={"original": "false"}, data=None
         )
-        
+
         # Verify the document was updated
         self.assertEqual(doc.content, b"file content")
         self.assertEqual(doc.content_type, "application/pdf")
@@ -97,12 +97,12 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_load_preview_mode(self):
         """
         Written By claude
-        
+
         Test loading a document in preview mode.
         """
         # Create a mock document
         doc = DownloadedDocument(id=1, mode=RetrieveFileMode.PREVIEW, original=True)
-        
+
         # Mock the response
         mock_response = MagicMock()
         mock_response.content = b"preview content"
@@ -111,15 +111,15 @@ class TestDownloadedDocumentResource(unittest.TestCase):
             "Content-Disposition": 'inline; filename="preview.pdf"',
         }
         self.mock_client.request_raw.return_value = mock_response
-        
+
         # Call the method
         self.resource.load(doc)
-        
+
         # Verify the client was called correctly
         self.mock_client.request_raw.assert_called_once_with(
             "GET", URLS.preview, params={"original": "true"}, data=None
         )
-        
+
         # Verify the document was updated
         self.assertEqual(doc.content, b"preview content")
         self.assertEqual(doc.content_type, "application/pdf")
@@ -129,12 +129,12 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_load_thumbnail_mode(self):
         """
         Written By claude
-        
+
         Test loading a document in thumbnail mode.
         """
         # Create a mock document
         doc = DownloadedDocument(id=1, mode=RetrieveFileMode.THUMBNAIL, original=False)
-        
+
         # Mock the response
         mock_response = MagicMock()
         mock_response.content = b"thumbnail content"
@@ -143,15 +143,15 @@ class TestDownloadedDocumentResource(unittest.TestCase):
             "Content-Disposition": 'inline; filename="thumb.jpg"',
         }
         self.mock_client.request_raw.return_value = mock_response
-        
+
         # Call the method
         self.resource.load(doc)
-        
+
         # Verify the client was called correctly
         self.mock_client.request_raw.assert_called_once_with(
             "GET", URLS.thumbnail, params={"original": "false"}, data=None
         )
-        
+
         # Verify the document was updated
         self.assertEqual(doc.content, b"thumbnail content")
         self.assertEqual(doc.content_type, "image/jpeg")
@@ -161,12 +161,12 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_load_default_mode(self):
         """
         Written By claude
-        
+
         Test loading a document with no mode specified (should default to download).
         """
         # Create a mock document with no mode
         doc = DownloadedDocument(id=1, mode=None, original=False)
-        
+
         # Mock the response
         mock_response = MagicMock()
         mock_response.content = b"file content"
@@ -175,10 +175,10 @@ class TestDownloadedDocumentResource(unittest.TestCase):
             "Content-Disposition": 'attachment; filename="test.pdf"',
         }
         self.mock_client.request_raw.return_value = mock_response
-        
+
         # Call the method
         self.resource.load(doc)
-        
+
         # Verify the client was called correctly with default mode (download)
         self.mock_client.request_raw.assert_called_once_with(
             "GET", URLS.download, params={"original": "false"}, data=None
@@ -187,12 +187,12 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_load_no_content_disposition(self):
         """
         Written By claude
-        
+
         Test loading a document when the response has no Content-Disposition header.
         """
         # Create a mock document
         doc = DownloadedDocument(id=1, mode=RetrieveFileMode.DOWNLOAD, original=False)
-        
+
         # Mock the response with no Content-Disposition
         mock_response = MagicMock()
         mock_response.content = b"file content"
@@ -201,10 +201,10 @@ class TestDownloadedDocumentResource(unittest.TestCase):
             # No Content-Disposition header
         }
         self.mock_client.request_raw.return_value = mock_response
-        
+
         # Call the method
         self.resource.load(doc)
-        
+
         # Verify the document was updated correctly
         self.assertEqual(doc.content, b"file content")
         self.assertEqual(doc.content_type, "application/pdf")
@@ -214,12 +214,12 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_load_complex_content_disposition(self):
         """
         Written By claude
-        
+
         Test loading a document with a complex Content-Disposition header.
         """
         # Create a mock document
         doc = DownloadedDocument(id=1, mode=RetrieveFileMode.DOWNLOAD, original=False)
-        
+
         # Mock the response with a complex Content-Disposition
         mock_response = MagicMock()
         mock_response.content = b"file content"
@@ -228,10 +228,10 @@ class TestDownloadedDocumentResource(unittest.TestCase):
             "Content-Disposition": 'attachment; filename="test file.pdf"; size=12345',
         }
         self.mock_client.request_raw.return_value = mock_response
-        
+
         # Call the method
         self.resource.load(doc)
-        
+
         # Verify the document was updated correctly
         self.assertEqual(doc.content, b"file content")
         self.assertEqual(doc.disposition_filename, "test file.pdf")
@@ -240,15 +240,15 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_load_request_failure(self):
         """
         Written By claude
-        
+
         Test that a ResourceNotFoundError is raised when the request fails.
         """
         # Create a mock document
         doc = DownloadedDocument(id=1, mode=RetrieveFileMode.DOWNLOAD, original=False)
-        
+
         # Mock the response to be None (request failed)
         self.mock_client.request_raw.return_value = None
-        
+
         # Verify that the correct exception is raised
         with self.assertRaises(ResourceNotFoundError):
             self.resource.load(doc)
@@ -257,12 +257,12 @@ class TestDownloadedDocumentResource(unittest.TestCase):
     def test_update_locally_called_correctly(self, mock_update_locally):
         """
         Written By claude
-        
+
         Test that update_locally is called with the correct parameters.
         """
         # Create a mock document
         doc = DownloadedDocument(id=1, mode=RetrieveFileMode.DOWNLOAD, original=False)
-        
+
         # Mock the response
         mock_response = MagicMock()
         mock_response.content = b"file content"
@@ -271,10 +271,10 @@ class TestDownloadedDocumentResource(unittest.TestCase):
             "Content-Disposition": 'attachment; filename="test.pdf"',
         }
         self.mock_client.request_raw.return_value = mock_response
-        
+
         # Call the method
         self.resource.load(doc)
-        
+
         # Verify update_locally was called with the correct parameters
         mock_update_locally.assert_called_once_with(
             from_db=True,
